+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Fix a typo in the README
+- Panic when trying to build a `Fenrir` instance without the required parameters set
 
 ## 0.4.1 - 2023-06-05